backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
numpy
scikit-learn
joblib
pandas
groq
pymongo